+++ src/DirectSolver/DirectSolver-CSR-LU-Take/directSolverTake.cpp
@@ -1,6 +1,6 @@
-#include "../../../include/DirectSolver/DirectSolverGiveCustomLU/directSolverGiveCustomLU.h"
+#include "../../../include/DirectSolver/DirectSolver-CSR-LU-Take/directSolverTakeCustomLU.h"
 
-DirectSolverGiveCustomLU::DirectSolverGiveCustomLU(const PolarGrid& grid, const LevelCache& level_cache,
+DirectSolverTakeCustomLU::DirectSolverTakeCustomLU(const PolarGrid& grid, const LevelCache& level_cache,
                                                    const DomainGeometry& domain_geometry,
                                                    const DensityProfileCoefficients& density_profile_coefficients,
                                                    bool DirBC_Interior, int num_omp_threads)
@@ -10,11 +10,11 @@ DirectSolverGiveCustomLU::DirectSolverGiveCustomLU(const PolarGrid& grid, const 
     lu_solver_     = SparseLUSolver<double>(solver_matrix_);
 }
 
-void DirectSolverGiveCustomLU::solveInPlace(Vector<double>& solution)
+void DirectSolverTakeCustomLU::solveInPlace(Vector<double>& solution)
 {
     lu_solver_.solveInPlace(solution);
 }
 
-DirectSolverGiveCustomLU::~DirectSolverGiveCustomLU()
+DirectSolverTakeCustomLU::~DirectSolverTakeCustomLU()
 {
 }
