+++ src/Level/level.cpp
@@ -3,10 +3,10 @@
 #include "../../include/Residual/ResidualGive/residualGive.h"
 #include "../../include/Residual/ResidualTake/residualTake.h"
 
-#include "../../include/DirectSolver/DirectSolverGive/directSolverGive.h"
-#include "../../include/DirectSolver/DirectSolverTake/directSolverTake.h"
-#include "../../include/DirectSolver/DirectSolverGiveCustomLU/directSolverGiveCustomLU.h"
-#include "../../include/DirectSolver/DirectSolverTakeCustomLU/directSolverTakeCustomLU.h"
+#include "../../include/DirectSolver/DirectSolver-COO-MUMPS-Give/directSolverGive.h"
+#include "../../include/DirectSolver/DirectSolver-COO-MUMPS-Take/directSolverTake.h"
+#include "../../include/DirectSolver/DirectSolver-CSR-LU-Give/directSolverGiveCustomLU.h"
+#include "../../include/DirectSolver/DirectSolver-CSR-LU-Take/directSolverTakeCustomLU.h"
 
 #include "../../include/Smoother/SmootherGive/smootherGive.h"
 #include "../../include/Smoother/SmootherTake/smootherTake.h"
